# Conflicts:
#   backend/build.gradle
#   backend/src/main/java/com/ssafy/cheertogether/member/dto/MemberJoinRequest.java

+++ backend/src/main/java/com/ssafy/cheertogether/member/dto/MemberModifyRequest.java
@@ -9,6 +9,6 @@ public class MemberModifyRequest {
 	private String nickname;
 	private String profileImage;
 	private String myInfo;
-	private List<Integer> favoriteLeagueList;
-	private List<Integer> favoriteTeamList;
+	private List<Long> favoriteLeagueList;
+	private List<Long> favoriteTeamList;
 }
